[Feat] Adds vinyl record condition field
Adds a condition field to the vinyl record entity.

This change introduces a new "condition" field, represented as an enum, to the vinyl record model.
The new field allows users to specify the condition of their vinyl records, like 'mint', 'near_mint', 'very_good', 'good', 'fair', and 'poor'.
The client form is also updated to allow for the user to set a condition when creating a vinyl record.